Added object modifications tracking using new Update API, for used with "Partial PLY export".
Objects are added to a modified list, which is checked during export. If the entire export is successful the object is removed from the modified list and added to a list of known exported objects.

The Update API will only track visible changes, thus will not detect if say "Render" level in the subdiv modifier is changed.
Also according to Brecht it may not be 100% reliable otherwise either.

http://wiki.blender.org/index.php/Dev:2.6/Source/Render/UpdateAPI