Fix issue with motion blur and animation of parent object.
If the parent object was animated but not the child object, then the child object would not get motion blur.

As per suggestion from Campbell Barton, we should use scene.set_frame() and access the object's matrices instead of manually constructing the sub-frame matrix from the fcurves, as this will respect drivers and similar as well.

Since cameras can handle scaling components in the transformation, the ignore_scale parameter has been removed.